+++ src/components/TipOfTheDay/TipOfTheDay.tsx
@@ -1,0 +1,17 @@
+const TIPS = [
+  {
+    body: 'Do you want to display legend separately? Use `renderHeatmapTrackerLegend(this.container, trackerData)` to generate it.'
+  },
+  {
+    body: 'Use `separateMonths` to separate months on the view! It looks great!'
+  },
+  {
+    body: 'You can enable/disable tip of the day in the settings'
+  },
+  {
+    body: 'To be honest, it is not the tip of the day. It selects a random tip every time you open Heatmap Tracker :)'
+  },
+  {
+    body: 'Check settings, there are some really nice features!',
+  }
+];